pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
    "pyside6-essentials~=6.7",
]
build-backend = "setuptools.build_meta"

[project]
name="splitguides"
description="Speedrun notes tool for advancing notes automatically with Livesplit."
authors = [
  { name="David C Ellis" },
]
readme="README.md"
requires-python = ">=3.12"  # 3.12 should function and build now
dependencies = [
    "pyside6~=6.7",
    "jinja2~=3.1",
    "bleach[css]==6.0",  # Each upgrade to bleach has broken something so pin it.
    "flask~=3.0",
    "markdown~=3.6",
    "keyboard~=0.13.5",
    "ducktools-classbuilder>=0.2.1",  # renamed in_dict to serialize
    "ducktools-env>=0.1.5",  # Used for script runner implementation
    "waitress~=3.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dynamic = ['version']
license = {file = "COPYING"}

[project.optional-dependencies]
testing = ["pytest", "pytest-cov", "pytest-qt"]
build = ["cx-freeze", "pywin32"]

[tool.setuptools.dynamic]
version = { attr = "splitguides.__version__" }

[tool.pytest.ini_options]
addopts= "--cov=src/ --cov-report=term-missing"
testpaths = [
    "tests",
]